AI/Chatbot/requirement.txt
google-generativeai==<version_number>
python-dotenv==1.0.0